+++ src/cirrus/lib/payload_bucket.py
@@ -1,43 +1,48 @@
+from __future__ import annotations
+
 import contextlib
 import os
 import uuid
 
-from typing import Any
+from typing import Any, Self
 
 from boto3utils import s3
 
 from cirrus.lib.errors import NoUrlError, UndefinedPayloadBucketError
 from cirrus.lib.utils import payload_from_s3
 
-PREFIX_ROOT = "cirrus"
-
-PREFIX_TMP = f"{PREFIX_ROOT}/tmp"  # noqa: S108
-PREFIX_BATCH = f"{PREFIX_TMP}/batch"
-PREFIX_INVALID = f"{PREFIX_TMP}/invalid"
-PREFIX_OVERSIZED = f"{PREFIX_TMP}/oversized"
-
-PREFIX_EXECS = f"{PREFIX_ROOT}/executions"
+DEFAULT_ROOT_PREFIX = "cirrus"
 
 INPUT_KEY = "input.json"
 OUTPUT_KEY = "output.json"
 
 
 class PayloadBucket:
-    def __init__(self, bucket_name: str | None = None) -> None:
-        self._bucket_name = bucket_name
+    def __init__(
+        self,
+        bucket_name: str,
+        root_prefix: str | None = None,
+    ) -> None:
+        self.bucket_name = bucket_name
+        self.root_prefix = (
+            root_prefix if root_prefix is not None else DEFAULT_ROOT_PREFIX
+        )
 
-    def name(self) -> str:
-        if self._bucket_name:
-            return self._bucket_name
+        self.prefix_tmp = f"{self.root_prefix}/tmp"  ## noqa: S108
+        self.prefix_batch = f"{self.prefix_tmp}/batch"
+        self.prefix_invalid = f"{self.prefix_tmp}/invalid"
+        self.prefix_oversized = f"{self.prefix_tmp}/oversized"
+        self.prefix_execs = f"{self.root_prefix}/executions"
 
-        payload_bucket = os.getenv("CIRRUS_PAYLOAD_BUCKET")
-
-        if not payload_bucket:
+    @classmethod
+    def from_env(cls) -> Self:
+        bucket_name = os.getenv("CIRRUS_PAYLOAD_BUCKET")
+        if not bucket_name:
             raise UndefinedPayloadBucketError(
                 "CIRRUS_PAYLOAD_BUCKET env var is not defined",
             )
-
-        return payload_bucket
+        root_prefix = os.getenv("CIRRUS_PAYLOAD_ROOT_PREFIX", DEFAULT_ROOT_PREFIX)
+        return cls(bucket_name, root_prefix)
 
     @staticmethod
     def parse_url(url: str) -> tuple[str, str]:
@@ -62,7 +67,7 @@ class PayloadBucket:
 
         prefix = prefix + "/" if prefix else prefix
 
-        url = f"s3://{self.name()}/{prefix}{key}"
+        url = f"s3://{self.bucket_name}/{prefix}{key}"
         s3().upload_json(payload, url)
         return url
 
@@ -73,7 +78,7 @@ class PayloadBucket:
         return self._upload_payload(
             payload=payload,
             key=f"{uuid.uuid1()}.json",
-            prefix=PREFIX_OVERSIZED,
+            prefix=self.prefix_oversized,
         )
 
     def upload_batch_payload(
@@ -85,7 +90,7 @@ class PayloadBucket:
             # ideally we'd use {uuid.uuid1()}/input.json but we keep this
             # format for backwards compatibility with the old cirrus.lib ways
             key=f"{payload['id']}/{uuid.uuid1()}.json",
-            prefix=PREFIX_BATCH,
+            prefix=self.prefix_batch,
             # in the batch case if we already have a url then we can just use it
             copy_from_url=False,
         )
@@ -97,16 +102,15 @@ class PayloadBucket:
         return self._upload_payload(
             payload=payload,
             key=f"{uuid.uuid1()}.json",
-            prefix=PREFIX_INVALID,
+            prefix=self.prefix_invalid,
         )
 
-    @classmethod
     def exec_payload_prefix(
-        cls,
+        self,
         payload_id: str,
         execution_id: str,
     ) -> str:
-        return f"{PREFIX_EXECS}/{payload_id}/{execution_id}"
+        return f"{self.prefix_execs}/{payload_id}/{execution_id}"
 
     def get_input_payload_url(
         self,
@@ -114,7 +118,7 @@ class PayloadBucket:
         execution_id: str,
     ) -> str:
         prefix = self.exec_payload_prefix(payload_id, execution_id)
-        return f"s3://{self.name()}/{prefix}/{INPUT_KEY}"
+        return f"s3://{self.bucket_name}/{prefix}/{INPUT_KEY}"
 
     def get_output_payload_url(
         self,
@@ -122,7 +126,7 @@ class PayloadBucket:
         execution_id: str,
     ) -> str:
         prefix = self.exec_payload_prefix(payload_id, execution_id)
-        return f"s3://{self.name()}/{prefix}/{OUTPUT_KEY}"
+        return f"s3://{self.bucket_name}/{prefix}/{OUTPUT_KEY}"
 
     def upload_input_payload(
         self,