+++ tests/lambda_functions/conftest.py
@@ -23,11 +23,11 @@ def publish_topic(sns):
 
 
 @pytest.fixture(autouse=True)
-def _env(_environment, queue, publish_topic, eventdb, statedb, payloads):
+def _env(_environment, queue, publish_topic, eventdb, statedb, payload_bucket):
     os.environ["CIRRUS_PROCESS_QUEUE_URL"] = queue["QueueUrl"]
     os.environ["CIRRUS_PUBLISH_TOPIC_ARN"] = publish_topic
     os.environ["CIRRUS_STATE_DB"] = statedb.table_name
     os.environ["CIRRUS_EVENT_DB_AND_TABLE"] = (
         f"{eventdb.event_db_name}|{eventdb.event_table_name}"
     )
-    os.environ["CIRRUS_PAYLOAD_BUCKET"] = payloads
+    os.environ["CIRRUS_PAYLOAD_BUCKET"] = payload_bucket.bucket_name